Crate google_jobs3

This documentation was generated from Cloud Talent Solution crate version 1.0.9+20190702, where 20190702 is the exact revision of the jobs:v3 schema built by the mako code generator v1.0.9.

Everything else about the Cloud Talent Solution v3 API can be found at the official documentation site. The original source code is on github.

Features

Handle the following Resources with ease from the central hub ...

projects

client events create, companies create, companies delete, companies get, companies list, companies patch, complete, jobs batch delete, jobs create, jobs delete, jobs get, jobs list, jobs patch, jobs search and jobs search for alert

Not what you are looking for ? Find all other Google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

creates Method Builders which in turn allow access to individual Call Builders

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

All structures are marked with applicable traits to further categorize them and ease browsing.

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.projects().jobs_create(...).doit()
let r = hub.projects().jobs_patch(...).doit()
let r = hub.projects().jobs_get(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-jobs3 = "*"
# This project intentionally uses an old version of Hyper. See
# https://github.com/Byron/google-apis-rs/issues/173 for more
# information.
hyper = "^0.10"
hyper-rustls = "^0.6"
serde = "^1.0"
serde_json = "^1.0"
yup-oauth2 = "^1.0"

A complete example

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_jobs3 as jobs3;
use jobs3::CreateJobRequest;
use jobs3::{Result, Error};
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use jobs3::CloudTalentSolution;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())),
                              <MemoryStorage as Default>::default(), None);
let mut hub = CloudTalentSolution::new(hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())), auth);
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req = CreateJobRequest::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.projects().jobs_create(req, "parent")
             .doit();
 
match result {
    Err(e) => match e {
        // The Error enum provides details about what exactly happened.
        // You can also just use its `Debug`, `Display` or `Error` traits
         Error::HttpError(_)
        |Error::MissingAPIKey
        |Error::MissingToken(_)
        |Error::Cancelled
        |Error::UploadSizeLimitExceeded(_, _)
        |Error::Failure(_)
        |Error::BadRequest(_)
        |Error::FieldClash(_)
        |Error::JsonDecodeError(_, _) => println!("{}", e),
    },
    Ok(res) => println!("Success: {:?}", res),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

Uploads and Downloads

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

Customization and Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

Optional Parts in Server-Requests

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifiable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are moved

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Structs

ApplicationInfo
Application related details of a job posting.

BatchDeleteJobsRequest
Input only.

BucketRange
Represents starting and ending value of a range in double.

BucketizedCount
Represents count of jobs within one bucket.

Chunk

ClientEvent
An event issued when an end user interacts with the application that implements Cloud Talent Solution. Providing this information improves the quality of search and recommendation for the API clients, enabling the service to perform optimally. The number of events sent must be consistent with other calls, such as job searches, issued to the service by the client.

CloudTalentSolution
Central instance to access all CloudTalentSolution related resource activities

CommuteFilter
Input only.

CommuteInfo
Output only.

Company
A Company resource represents a company in the service. A company is the entity that owns job postings, that is, the hiring entity responsible for employing applicants for the job position.

CompanyDerivedInfo
Derived details about the company.

CompensationEntry
A compensation entry that represents one component of compensation, such as base pay, bonus, or other compensation type.

CompensationFilter
Input only.

CompensationHistogramRequest
Input only.

CompensationHistogramResult
Output only.

CompensationInfo
Job compensation details.

CompensationRange
Compensation range.

CompleteQueryResponse
Output only.

CompletionResult
Output only.

ContentRange
Implements the Content-Range header, for serialization only

CreateClientEventRequest
The report event request.

CreateCompanyRequest
Input only.

CreateJobRequest
Input only.

CustomAttribute
Custom attribute values that are either filterable or non-filterable.

CustomAttributeHistogramRequest
Custom attributes histogram request. An error is thrown if neither string_value_histogram or long_value_histogram_bucketing_option has been defined.

CustomAttributeHistogramResult
Output only.

DefaultDelegate
A delegate with a conservative default implementation, which is used if no other delegate is set.

DeviceInfo
Device information collected from the job seeker, candidate, or other entity conducting the job search. Providing this information improves the quality of the search results across devices.

DummyNetworkStream

Empty
A generic empty message that you can re-use to avoid defining duplicated empty messages in your APIs. A typical example is to use it as the request or the response type of an API method. For instance:

ErrorResponse
A utility to represent detailed errors we might see in case there are BadRequests. The latter happen if the sent parameters or request structures are unsound

HistogramFacets
Input only.

HistogramResult
Output only.

HistogramResults
Output only.

Job
A Job resource represents a job posting (also referred to as a "job listing" or "job requisition"). A job belongs to a Company, which is the hiring entity responsible for the job.

JobDerivedInfo
Output only.

JobEvent
An event issued when a job seeker interacts with the application that implements Cloud Talent Solution.

JobQuery
Input only.

JsonServerError
A utility type which can decode a server response that indicates error

LatLng
An object representing a latitude/longitude pair. This is expressed as a pair of doubles representing degrees latitude and degrees longitude. Unless specified otherwise, this must conform to the WGS84 standard. Values must be within normalized ranges.

ListCompaniesResponse
Output only.

ListJobsResponse
Output only.

Location
Output only.

LocationFilter
Input only.

MatchingJob
Output only.

MethodInfo
Contains information about an API request.

Money
Represents an amount of money with its currency type.

MultiPartReader
Provides a Read interface that converts multiple parts into the protocol identified by RFC2387. Note: This implementation is just as rich as it needs to be to perform uploads to google APIs, and might not be a fully-featured implementation.

NumericBucketingOption
Input only.

NumericBucketingResult
Output only.

PostalAddress
Represents a postal address, e.g. for postal delivery or payments addresses. Given a postal address, a postal service can deliver items to a premise, P.O. Box or similar. It is not intended to model geographical locations (roads, towns, mountains).

ProcessingOptions
Input only.

ProjectClientEventCreateCall
Report events issued when end user interacts with customer's application that uses Cloud Talent Solution. You may inspect the created events in self service tools. Learn more about self service tools.

ProjectCompanyCreateCall
Creates a new company entity.

ProjectCompanyDeleteCall
Deletes specified company. Prerequisite: The company has no jobs associated with it.

ProjectCompanyGetCall
Retrieves specified company.

ProjectCompanyListCall
Lists all companies associated with the service account.

ProjectCompanyPatchCall
Updates specified company. Company names can't be updated. To update a company name, delete the company and all jobs associated with it, and only then re-create them.

ProjectCompleteCall
Completes the specified prefix with keyword suggestions. Intended for use by a job search auto-complete search box.

ProjectJobBatchDeleteCall
Deletes a list of Jobs by filter.

ProjectJobCreateCall
Creates a new job.

ProjectJobDeleteCall
Deletes the specified job.

ProjectJobGetCall
Retrieves the specified job, whose status is OPEN or recently EXPIRED within the last 90 days.

ProjectJobListCall
Lists jobs by filter.

ProjectJobPatchCall
Updates specified job.

ProjectJobSearchCall
Searches for jobs using the provided SearchJobsRequest.

ProjectJobSearchForAlertCall
Searches for jobs using the provided SearchJobsRequest.

ProjectMethods
A builder providing access to all methods supported on project resources. It is not used directly, but through the CloudTalentSolution hub.

RangeResponseHeader

RequestMetadata
Input only.

ResponseMetadata
Output only.

ResumableUploadHelper
A utility type to perform a resumable upload from start to end.

SearchJobsRequest
Input only.

SearchJobsResponse
Output only.

ServerError

ServerMessage

SpellingCorrection
Output only.

TimeOfDay
Represents a time of day. The date and time zone are either not significant or are specified elsewhere. An API may choose to allow leap seconds. Related types are google.type.Date and google.protobuf.Timestamp.

TimestampRange
Message representing a period of time between two timestamps.

UpdateCompanyRequest
Input only.

UpdateJobRequest
Input only.

XUploadContentType
The X-Upload-Content-Type header.

Enums

Error

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.

Traits

CallBuilder
Identifies types which represent builders for a particular resource method

Delegate
A trait specifying functionality to help controlling any request performed by the API. The trait has a conservative default implementation.

Hub
Identifies the Hub. There is only one per library, this trait is supposed to make intended use more explicit. The hub allows to access all resource methods more easily.

MethodsBuilder
Identifies types for building methods of a particular resource type

NestedType
Identifies types which are only used by other types internally. They have no special meaning, this trait just marks them for completeness.

Part
Identifies types which are only used as part of other types, which usually are carrying the Resource trait.

ReadSeek
A utility to specify reader types which provide seeking capabilities too

RequestValue
Identifies types which are used in API requests.

Resource
Identifies types which can be inserted and deleted. Types with this trait are most commonly used by clients of this API.

ResponseResult
Identifies types which are used in API responses.

ToParts
A trait for all types that can convert themselves into a parts string

UnusedType
Identifies types which are not actually used by the API This might be a bug within the google API schema.

Functions

remove_json_null_values

Type Definitions

Result
A universal result type used as return for all calls.