Struct google_jobs3::api::SearchJobsRequest

pub struct SearchJobsRequest {
    pub disable_keyword_match: Option<bool>,
    pub diversification_level: Option<String>,
    pub enable_broadening: Option<bool>,
    pub histogram_facets: Option<HistogramFacets>,
    pub job_query: Option<JobQuery>,
    pub job_view: Option<String>,
    pub offset: Option<i32>,
    pub order_by: Option<String>,
    pub page_size: Option<i32>,
    pub page_token: Option<String>,
    pub request_metadata: Option<RequestMetadata>,
    pub require_precise_result_size: Option<bool>,
    pub search_mode: Option<String>,
}

Input only. The Request body of the SearchJobs call.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

jobs search projects (request)

jobs search for alert projects (request)

Fields
disable_keyword_match: Option<bool>

Optional. Controls whether to disable exact keyword match on Job.job_title, Job.description, Job.company_display_name, Job.locations, Job.qualifications. When disable keyword match is turned off, a keyword match returns jobs that do not match given category filters when there are matching keywords. For example, the query "program manager," a result is returned even if the job posting has the title "software developer," which does not fall into "program manager" ontology, but does have "program manager" appearing in its description. For queries like "cloud" that does not contain title or location specific ontology, jobs with "cloud" keyword matches are returned regardless of this flag's value. Please use Company.keyword_searchable_custom_fields or Company.keyword_searchable_custom_attributes if company specific globally matched custom field/attribute string values is needed. Enabling keyword match improves recall of subsequent search requests. Defaults to false.

diversification_level: Option<String>

Optional. Controls whether highly similar jobs are returned next to each other in the search results. Jobs are identified as highly similar based on their titles, job categories, and locations. Highly similar results are clustered so that only one representative job of the cluster is displayed to the job seeker higher up in the results, with the other jobs being displayed lower down in the results. Defaults to DiversificationLevel.SIMPLE if no value is specified.

enable_broadening: Option<bool>

Optional. Controls whether to broaden the search when it produces sparse results. Broadened queries append results to the end of the matching results list. Defaults to false.

histogram_facets: Option<HistogramFacets>

Optional. Histogram requests for jobs matching JobQuery.

job_query: Option<JobQuery>

Optional. Query used to search against jobs, such as keyword, location filters, etc.

job_view: Option<String>

Optional. The desired job attributes returned for jobs in the search response. Defaults to JobView.SMALL if no value is specified.

offset: Option<i32>

Optional. An integer that specifies the current offset (that is, starting result location, amongst the jobs deemed by the API as relevant) in search results. This field is only considered if page_token is unset. The maximum allowed value is 5000. Otherwise an error is thrown. For example, 0 means to return results starting from the first matching job, and 10 means to return from the 11th job. This can be used for pagination, (for example, pageSize = 10 and offset = 10 means to return from the second page).

order_by: Option<String>

Optional. The criteria determining how search results are sorted. Default is "relevance desc". Supported options are: * "relevance desc": By relevance descending, as determined by the API algorithms. Relevance thresholding of query results is only available with this ordering. * "posting_publish_time desc": By Job.posting_publish_time descending. * "posting_update_time desc": By Job.posting_update_time descending. * "title": By Job.title ascending. * "title desc": By Job.title descending. * "annualized_base_compensation": By job's CompensationInfo.annualized_base_compensation_range ascending. Jobs whose annualized base compensation is unspecified are put at the end of search results. * "annualized_base_compensation desc": By job's CompensationInfo.annualized_base_compensation_range descending. Jobs whose annualized base compensation is unspecified are put at the end of search results. * "annualized_total_compensation": By job's CompensationInfo.annualized_total_compensation_range ascending. Jobs whose annualized base compensation is unspecified are put at the end of search results. * "annualized_total_compensation desc": By job's CompensationInfo.annualized_total_compensation_range descending. Jobs whose annualized base compensation is unspecified are put at the end of search results.

page_size: Option<i32>

Optional. A limit on the number of jobs returned in the search results. Increasing this value above the default value of 10 can increase search response time. The value can be between 1 and 100.

page_token: Option<String>

Optional. The token specifying the current offset within search results. See SearchJobsResponse.next_page_token for an explanation of how to obtain the next set of query results.

request_metadata: Option<RequestMetadata>

Required. The meta information collected about the job searcher, used to improve the search quality of the service. The identifiers (such as user_id) are provided by users, and must be unique and consistent.

require_precise_result_size: Option<bool>

This field is deprecated.

search_mode: Option<String>

Optional. Mode of a search. Defaults to SearchMode.JOB_SEARCH.